Extract atomicWriteJsonSync to shared @/lib/paths utility
Move duplicated atomicWriteJsonSync function from three route files
into the shared paths.ts module and import it from there.
